requirements.txt
asgiref==3.6.0
Django==4.1.4
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-environ==0.9.0
django-mathfilters==1.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.1.1
Pillow==9.4.0
PyJWT==2.6.0
pytz==2022.7
sqlparse==0.4.3
tzdata==2022.7
